+++ src/main/java/components/control/NumberBoxComponent.java
@@ -228,7 +228,7 @@ public class NumberBoxComponent extends ACtrlComponent {
                 case KeyEvent.VK_ENTER:
                     fireEventAdjustmentBegin();
                     try {
-                        setValue(Float.parseFloat(keybBuffer));
+                        setValue(Double.parseDouble(keybBuffer));
                     } catch (java.lang.NumberFormatException ex) {
                     }
                     fireEventAdjustmentFinished();
